move field to me
