+++ src/schema/v2/me/index.ts
@@ -38,6 +38,7 @@ import LotStanding from "./lot_standing"
 import LotStandings from "./lot_standings"
 import { MyBids } from "./myBids"
 import { MyCollection } from "./myCollection"
+import newWorksByInterestingArtists from "./new_works_by_interesting_artists"
 import { RecentlyViewedArtworks } from "./recently_viewed_artworks"
 import { SaleRegistrationConnection } from "./sale_registrations"
 import { SavedArtworks } from "./savedArtworks"
@@ -135,6 +136,7 @@ const Me = new GraphQLObjectType<any, ResolverContext>({
     name: {
       type: GraphQLString,
     },
+    newWorksByInterestingArtists: newWorksByInterestingArtists,
     initials: initials("name"),
     paddleNumber: {
       type: GraphQLString,
